// react/components/api/checkoutexternal/models/InvoiceStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schematic API
 * Schematic API
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const InvoiceStatus = {
    Draft: 'draft',
    Open: 'open',
    Paid: 'paid',
    Uncollectible: 'uncollectible',
    Void: 'void'
} as const;
export type InvoiceStatus = typeof InvoiceStatus[keyof typeof InvoiceStatus];


export function instanceOfInvoiceStatus(value: any): boolean {
    for (const key in InvoiceStatus) {
        if (Object.prototype.hasOwnProperty.call(InvoiceStatus, key)) {
            if (InvoiceStatus[key as keyof typeof InvoiceStatus] === value) {
                return true;
            }
        }
    }
    return false;
}

export function InvoiceStatusFromJSON(json: any): InvoiceStatus {
    return InvoiceStatusFromJSONTyped(json, false);
}

export function InvoiceStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): InvoiceStatus {
    return json as InvoiceStatus;
}

export function InvoiceStatusToJSON(value?: InvoiceStatus | null): any {
    return value as any;
}

export function InvoiceStatusToJSONTyped(value: any, ignoreDiscriminator: boolean): InvoiceStatus {
    return value as InvoiceStatus;
}